requirements.txt
Django==4.0
psycopg2-binary==2.9.9
djangorestframework==3.13.1
django-filter==21.1
pillow==9.3.0
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
django-cors-headers==4.3.0
tensorflow==2.14.0
pandas==2.0.3
joblib==1.3.1
python-dotenv==1.0.0
scikit-learn==1.3.0